refactor: move prompt policy into runtime
