+++ tests/data/known_runtime_telegram_bot_couplings.json
@@ -1,12 +1,4 @@
 {
-  "src/runtime/generation/service.py": [
-    "telegram_bot.integrations.prompt_manager",
-    "telegram_bot.integrations.prompt_templates",
-    "telegram_bot.observability",
-    "telegram_bot.services.coverage_mode",
-    "telegram_bot.services.metrics",
-    "telegram_bot.services.response_style_detector"
-  ],
   "src/runtime/graph/builder.py": [
     "telegram_bot.graph.graph:build_graph"
   ],
@@ -23,8 +15,5 @@
     "telegram_bot.services.query_preprocessor",
     "telegram_bot.services.rag_core",
     "telegram_bot.services.small_to_big"
-  ],
-  "src/runtime/services/query_preprocessor.py": [
-    "telegram_bot.integrations.prompt_manager"
   ]
 }
